+++ react/MuiCozyTheme/ListItem/ListItemBase/index.jsx
@@ -9,9 +9,9 @@ import IconButton from '../../../IconButton'
 import Icon from '../../../Icon'
 import DotsIcon from '../../../Icons/Dots'
 import ExpandedAttributes from '../ExpandedAttributes'
-import ActionMenuWithClose from './ActionsMenu/ActionMenuWithClose'
+import ActionMenuWithClose from '../../../ActionMenu/ActionMenuWithClose'
 import { ActionMenuHeader } from '../../../ActionMenu'
-import ActionsItems from './ActionsMenu/ActionsItems'
+import ActionsItems from '../../../ActionMenu/Actions/ActionsItems'
 import useBreakpoints from '../../../hooks/useBreakpoints'
 import RenameInput from './Renaming/RenameInput'
 import withListItemLocales from '../hoc/withListItemLocales'
